Highlighting Code
Change tiling size to 25 from 35.

Wrote code to manage clicks so that it correctly corrolates the
clicking to an index of the grid.

Clicks now highlight blocks that can be driven on. This doesn't work
with dragging but that will be the next move. I need to also check if
when the drag is released the location is next to a fire, if not then
flash red and remove. If valid then flash green, create a firetruck and
set it's path then remove the highlighting.