Massive Frontend Updates
Includes a whole heap of new libraries to do scrolling in the homepage
which has also been commited with this.

There's still more work to be done on the development of the code.

There's also a bug to do with caching #values after an HTML page which
is causing scrolling to not work. Might have to use something like
slidejs instead.